+++ Source-Files-Folder/Engine-Logic-Folder/Header-Files-Folder/chess-board-values.h
@@ -2,26 +2,26 @@
 #ifndef CHESS_BOARD_VALUES_H
 #define CHESS_BOARD_VALUES_H
 
-extern const unsigned short BOARD_RANKS;
-extern const unsigned short BOARD_FILES;
-extern const unsigned short BOARD_LENGTH;
+extern const uint8_t BOARD_RANKS;
+extern const uint8_t BOARD_FILES;
+extern const uint8_t BOARD_LENGTH;
 
-extern const unsigned short PIECE_TYPE_SPAN;
-extern const unsigned short PIECE_TEAM_SPAN;
+extern const uint8_t PIECE_TYPE_SPAN;
+extern const uint8_t PIECE_TEAM_SPAN;
 
-extern const signed short BLACK_MOVE_VALUE;
-extern const signed short WHITE_MOVE_VALUE;
+extern const int8_t BLACK_MOVE_VALUE;
+extern const int8_t WHITE_MOVE_VALUE;
 
-extern const unsigned short WHITE_START_RANK;
-extern const unsigned short BLACK_START_RANK;
+extern const uint8_t WHITE_START_RANK;
+extern const uint8_t BLACK_START_RANK;
 
-extern const unsigned short WHITE_PAWN_RANK;
-extern const unsigned short BLACK_PAWN_RANK;
+extern const uint8_t WHITE_PAWN_RANK;
+extern const uint8_t BLACK_PAWN_RANK;
 
-extern const unsigned short KING_START_FILE;
+extern const uint8_t KING_START_FILE;
 
-extern const signed short KSIDE_FILE_OFFSET;
-extern const signed short QSIDE_FILE_OFFSET;
+extern const int8_t KSIDE_FILE_OFFSET;
+extern const int8_t QSIDE_FILE_OFFSET;
 
 extern const Point WHITE_KING_POINT;
 extern const Point BLACK_KING_POINT;
