+++ Source-Files-Folder/Engine-Logic-Folder/Header-Files-Folder/engine-point-values.h
@@ -2,13 +2,13 @@
 #ifndef ENGINE_POINT_VALUES_H
 #define ENGINE_POINT_VALUES_H
 
-typedef signed short Point;
+typedef int8_t Point;
 
 extern const Point POINT_RANK_MASK;
 extern const Point POINT_FILE_MASK;
 
-extern const unsigned short POINT_RANK_SHIFT;
-extern const unsigned short POINT_FILE_SHIFT;
+extern const uint8_t POINT_RANK_SHIFT;
+extern const uint8_t POINT_FILE_SHIFT;
 
 extern const Point POINT_NONE;
 extern const Point POINT_BLANK;
